Update patch set 1
Patch Set 1:

The goal was to make the dialog non-modal and keep it on top of the parent workbench window. The title bar is only removed on GTK -- this is an SWT limitation on GTK when ON_TOP is used.

But in this case, the ON_TOP was a bad idea, since it causes the dialog to be on top of all windows. Just removing ON_TOP in all 4 places solves the problem (and still keeps the dialog on top of the main window, since that one is already the parent shell).

Patch-set: 1