Update patch set 6
Patch Set 6:

In deed, I've notice it but it was to late to "revert" this. I know that I've brokre the build and in deed I'm sorry about that :(

Now we can revert this change (thos isn't preffered choice) or we can wait for jgit change. Honestly I would peek up second solution, but this is only my humble opinion, I'm open for any further discussions, suggestions ...

Honestly, I'm not against reverting this change, but I see revert as a boiler plait entry in EGit history, because it'll finally get merged ...

Patch-set: 6
Label: SUBM=+1