Update patch set 1
Patch Set 1:

> I think I prefer Robin's proposal in change
> Ifb31dc38ea673b096869437b03e3c738fa89a07c to use JGit's git 
> timestamp formatter which supports all the standard git 
> timestamp formats so that we can later let the user choose > his preferred format

I still find it a bad idea that specific plug-ins introduce their own formatting perference. It results in a product which looks like patch work. Also, as an EGit user I would expect to see the date in the (EGit) History view rendered the same as when I e.g. open the commit viewer. There I see a perfectly nicely formatted German date as expected:
"Dani Megert <dmegert> on 27.10.11 12:50" and I would simply expect to see the same in the History view.

If we think that the special Git format is important, then a preference should be added that allows to toggle between normal user readable format and the Git format. If the user chooses the Git format, then all UI in EGit should use it, i.e. also in the Commit editor.

Also note that there is currently a preference to set the date format for the change set date formatting (bug 356052).

Patch-set: 1