Update patch set 1
Patch Set 1:

> It seems Java doesn't care about the OS settings. It just 
> uses the locale (from OS unless overridden) and Java then 
> as a a per-locale database within Java.

I can't speak for Mac but on Windows it respects the regional settings in the sense that it respects the general formatting rules for the selected language/country. It however does not respect the detailed setting, e.g. the format chosen for short and long style. This is coming from the JRE as defined by DateFormat SHORT, MEDIUM and LONG.


The patch is better than nothing but still not what I would expect to see as a EGit user. See my comment in http://egit.eclipse.org/r/#change,4426.

Patch-set: 1