Update patch set 1
Patch Set 1:

I don't have a Mac at hand. Could you put a breakpoint on
  copyItem.setEnabled(styledText.getSelectionCount() > 0)
and check whether this is called and that the selection count is? Such code is also used in the TextViewer. I would be surprised if this isn't working on the Mac.

Regarding the menu: I guess if we offer the two copy menu items, we can leave the 'Select All' away, right?

Patch-set: 1