Update patch set 1
Patch Set 1:

Mathias, we looked into it and are still not able to reproduce the problem using your steps. However, if we remove the focus from the commit SHA1 (e.g. by clicking into the message entry field) and then open the context menu, we can see the problem. It comes from calling setSelection(0) in the focus listener. Can you confirm that your given steps are missing a step that causes the focuseLost event?

Patch-set: 1