Remove Gerrit Configuration Page from Clone Wizard
Now we have the possibility to plug in clone sources. The Gerrit
configuration will be provided by the appropriate Gerrit extension.

If this extension is not installed, or the user has cloned from
another source or using a custom URI, he can still configure the
remote as Gerrit in the context menu of the Repositories View.

Change-Id: Ia7f6654447e699d0b2d785efd852652b0a470d69
Signed-off-by: Stefan Lay <stefan.lay@sap.com>