Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(6 inline comments)

Nice improvement :-)

This seems to always propagate the ? decorator on the working set node when IndexDiff isn't yet available or still being computed. Instead it should not propagate any git decoration until the IndexDiff data is available in order not to confuse the user.

If you want to show that some operation is ongoing you may consider to render the node's text with an italic font until the index diff computation is ready.

Patch-set: 1