Allow push of single Refs/Commits from Repo & History Views
This adds a new SimplePushWizard which just allows selecting a target
remote and ref name. It then pushes out the selected Ref/Commit to the
selected target ref.

This fixes three bugs (see below), and i feel it would be a really cool
feature for 1.3!

There is room for improvement, and i need help with implementing content
assist for the target ref selection - this should be a combo much like
the destination combo in the existing RefSpecPanel

Bug: 341076
Bug: 356314
Bug: 337747
Change-Id: If3be09886c23b031489116d448cae2dce29e281a