Support tracking project at root of repository
Previously a project at the root of the working directory
was not tracked since they didn't have a relative path and
were not stored correctly in the memento.

This case is now explicitly handled and a project at the root
of the repository's working directory will now be tracked and
re-opened when switching branches.

Change-Id: I6ad4620b866fc2bce0ee02e47782bc5a2271aa52