[historyView] Hide additional Refs by default
The normal user is not interested in Refs like FETCH_HEAD, ORIG_HEAD
etc. when using the HistoryView. However, these Refs can pollute the
view as they can not be hidden independently if "Show All Branches" is
active.

The additional Refs are now hidden by default and the user can have
them displayed by checking the Show->Additional Refs menu item toggle
button. The HEAD Ref is always shown independent from this switch.

The toggle was also added to the History Preference page (along with
the "Show all Branches and Tags" toggle).

When the user does an import from Gerrit and selects either "Update
FETCH_HEAD only" or "Checkout FETCH_HEAD" and the toggle is off,
a checkbox will be shown informing the user about this fact and adding
the possibility to change this toggle during the course of the fetch.

Change-Id: I2b5df21e88ab7015012ef670de33819bd65de037
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>