Create change
Patch Set 1:

Well, I *think* this may be a problem. The body of the message is the explanation.

It is not clear to me if we may work with a Document from
any thread in Eclipse. Perhaps a Docuement is far enough away from the UI. Let someone with insights into this comment.

Patch-set: 1
Change-id: I04b731abf0126087622292daa6f39600ade8f945
Subject: Postpone handling of ref changes in GitDocument to the event loop
Branch: refs/heads/master
Commit: 1da934c9dc25b4dac69044e854215ff9e0582f53