Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

rebased to adapt build to new target release 1.0.0

I like the idea 

It's not obvious from the dialog layout that the commit-ish edit field at the bottom is an alternative to selecting a branch in the tree at the top. Either the editable commit-ish field can be moved inside the tree or we need another way to visualize these alternatives.

Patch-set: 2