Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Looks good, I like where this is going!

- What happens at task deactivation?

by default and easiest solution would be to go back to master. We could add a preference and allow people to choose to go back to master always or the last branch you were on (we would have to remember this in EGit via a preference probably or even better, just check the git ref log)

Patch-set: 1