Update patch set 1
Patch Set 1:

> since its awkward to get the logic right when it isn't.

It is not just awkward, it is necessary to get it right and to support git repositories which are not under the container path.

As for the GUI view, there must be better ways to handle the truncation. One suggestion is to consider presenting a container relative path for cases where possible and the full path where not.

Patch-set: 1