Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

There is still another case where the new config parameter is not taken into account: If you Select "Create Branch..." in the "Branches"-Node or the "Branches/Local"-Node of the Repository View the currently checked out branch is taken.

The code is in CreateBranchCommand, line 64.

Patch-set: 3