Update patch set 2
Patch Set 2: No score

re 1: '>' stands for "dirty" and for me, as soon as there is any diff compared to what's in the repository, such a project is dirty. Otherwise you have to rename the label for '>' to "interesting changes". Also, the fact that EGit shows a '?' indicates that it is not completely unimportant.

re 2: I see that the information is redundant but keep in mind that for disabled people, the icon is not "readable", hence '>' is an important information for them.

Patch-set: 2