Update patch set 1
Patch Set 1:

@Mathias Kinzler
The action classes don't have to be public. The only class that must be accessible by other plugins is GitFileRevisionReference.java


@Matthias Sohn
For example: our plugin displays a list of changed files. We know that they can be part of EGit repository, but from our plugin we can't perform such actions as described here.
After this patch we just need to implement adapter to GitFileRevisionReference and the actions will appear in context menu.

Patch-set: 1