Update patch set 6
Patch Set 6: I would prefer that you didn't submit this

I have problems with this. I agree that the display of ORIG_HEAD, FETCH_HEAD and MERGE_HEAD should be avoided for the 99% of the cases. But I would say that they are nearly always important to be displayed when the point to objects which are otherwise unreferenced. My suggestion: By default always show them when they point to something which is not referenced by any other ref. And add a preference which allows me to show them always & never.

Here is my argument: Even middle-experienced EGit users start doing rebases quite soon. If I am in the middle of a rebase I sometimes want to compare the state before I started the rebase operation to the current state. Without ORIG_HEAD the state where I startet rebase is lost.

Another argument: Even the gerrit Web-UI suggest the usage of FETCH_HEAD. If I just want to look at a specific proposal I want to be able to inspect it locally without having to create and afterwards delete a temporary ref. FETCH_HEAD is perfect for that.

You may say: that's your use-case but 99% of the normal users don't work like that and are beeing disturbed by these technical refs. My argument: Fine, if they don't work like that then they will never have ORIG_HEAD,MERGE_HEAD and FETCH_HEAD pointing to something which is otherwise unreferenced. Means: they will not see these refs with the default setting I propose.

Patch-set: 6