Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I do not like the way the sharing dialog looks with this. The Repository Path the dialog (almost always) gets truncated.

I do agree with the idea of making the directory absolute in code, since its awkward to get the logic right when it isn't.

Patch-set: 1