Update patch set 1
Patch Set 1:

I have read the bug report, but don't see what the problem is. Of course there are name conflicts if you try to create a new project in the workspace with the same directory as the repository (duh). When the project is named the same as the repository, it's even now not possible to have another project named the same (Eclipse limitation). And the second point sounds like a bug in M2Eclipse. It seems to me the reporter didn't understand the difference between the workspace directory structure and projects. And all these points also apply to CVS and SVN.

If you delete the workspace, all directories within are also deleted, that should also obvious and is no different than when you have uncommitted changes in CVS or SVN checkouts. Also, if you delete the project in Eclipse including project contents on disk, you also lost your repo.

All I'm trying to say is this: Please reconsider the decision for the default. We are migrating from CVS to Git in our company and this behaviour irritated all people without exception as I was showing them how to use EGit. It will also irritate other as EGit adoption grows.

Patch-set: 1