[repoView] Add submodule add/sync/update support
The Submodules node and all child repository nodes now have
an Update and Sync action available from the context menu.

Root repositories and 'Submodules' nodes now have an Add action
that opens a wizard to configure a new submodule that will be
configured in the parent repository and cloned on completion.

Change-Id: I2afea7dca9e40a2748c10a7f835b2deae301ba73
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>