Show commit in history view when annotation is selected
BlameOperation now listens for revision selection events
and shows/selects the commit in the history view when a
revision is selected.

This changes the behavior to initially load the history view
while the editor ruler is being shown so it is ready to show
selections once the editor and gutter area are explored.

Bug: 356313
Change-Id: Ia9cf2c6ec44206ee3567367487df7f4a6f11cd42