Update patch set 5
Patch Set 5:

The duplicate logic is there in the case when you want multiple entries of the same key, such as fetch ref specs.

The suffix's given to duplicate entries ([0],[1], etc.) has always been there and are unchanged by this commit.

The duplicate button does the same thing as the bottom add button did previously when a value was already selected.

I could remove the duplicate button but then the user would not be able to add a second value for the same key since the 'New Entry...' dialog prevents that.

I was trying to retain all the use cases the previous design had for adding, editing, removing, and duplicating entries.

Patch-set: 5