Update patch set 6
Patch Set 6: I would prefer that you didn't submit this

I tested this with the fix for bug 372020 and also with an additional fix that I had to make in PDE (bug 372726). Overall it looks good, but when one imports an already existing project the patch causes havoc: it deletes the project on the file system (i.e. from the repository).

Note that this will be more tricky than just replacing the deletion with removing the link in the Git importer, since once cannot assume that the existing project was imported by/from the same importer/repository.

Patch-set: 6