Update patch set 1
Patch Set 1: (6 inline comments)

very likely you didn't miss anything. the refresh button does still call reload, which re-gets the diff from the cache - but if all goes well (actually there is not much room for not-going-well :)) it has no effect, as the listeners already updated it to the correct state. shall i remove the refresh button, or leave it as a "emergency button" if something goes out of sync?

new patchset ahead.

Patch-set: 1