Proposal: make the staging view use the IndexDiffCache
This makes the staging view no longer use a separat "Staging view
indexer job", but rather uses the already existant (and always updated)
IndexDiffCacheEntry for the selected repository.

This is WIP, especially as i can't get the SWTBot test to behave.
everything works if i click through manually, but when the bot selects
the repository node in the test, the staging view is not notified, and
thus the view is not updated (breaking the test).

all suggestions welcome!

Change-Id: Ia2ed34f57e4639fd086b65ce421dfc7613f93f57