Update patch set 2
Patch Set 2:

IMO we should keep the refresh button, it's handy to be able to explicitly trigger refresh e.g. after doing changes outside the workbench

Please update the commit message so that we can submit this change when Jens reviewed it.

On which platform do you see test failures ? Which test is failing ?

Patch-set: 2