Update patch set 2
Patch Set 2:

i see the test failure on my gentoo linux in a Xephyr nested x server, running the StagingViewTest solely. it fails because the selection of the repository node in the (inactive!) repository view in the background does not trigger a selection event. this causes the staging view to not do anything - and the test fails. it works when i manually activate the view and click on the repository node - no idea whats wrong with the test.

new patchset with new commit message is coming up...

Patch-set: 2