Update patch set 7
Patch Set 7: I would prefer that you didn't submit this

(3 inline comments)

A remark on the refresh action: it might happen that a cached IndexDiff is not up to data. There is  a method refresh in class IndexDiffCacheEntry.
I think refreshAction.run should just call IndexDiffCacheEntry.refresh to trigger a complete recalculation of the IndexDiff.

Patch-set: 7