Update patch set 1
Patch Set 1: IP review completed

Very nice :-)

When the repository is selected in the Repositories View (or any node under the repository) the icon in the git toolbar looks disabled even when actions "Abort" and "Continue" in the drop-down are enabled. I think this is a general problem and selection of nodes in Repositories View doesn't enable corresponding git actions in git toolbar but it should do so. Probably some adapter to RepositoryViewNode types is missing here.

I think this drop-down should be also added to the staging view as this is a very convenient place to find and resolve conflicts.

Patch-set: 1