Fix an occasional NPE in staging view with new repos
Sometimes when starting from a fresh workspace and creating a repo from
scratch, the staging view throws unhandled event loop exceptions because
repository is null in CommitMessageComponent. This can be avoided by
re-setting the repository also in the case where the staging view thinks
it didn't change. Since this only updates a field, the setting does not
harm whatsoever.

Change-Id: I0ae9f39c1e7a1338efeed5e03701ab4eb2e49c2a