Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Now after having resolved a conflict the user can press the button "Rebase" again which opens the branch selection dialog. This is not meaningful in state "Rebase interactive". Is it possible to disable the button but enabling the list in this case?

Patch-set: 1