[historyView] Track renamed paths and old paths to actions
Several History view actions currently fail when the view is
set to follow renames since they derive the path used from
the current repository path and do not take into account
older paths that the resource was known as.

This commit creates a simple RenameTracker class that tracks
which commits introduced renames allowing the GitHistoryPage
to provide the previous paths for a specific commit given the
initial path.

Bug: 374722
Change-Id: Ieddbfb06418b146e5066a7ef2eda51ab4af858e2