Allow to show/edit branch configuration
It should be possible to see/change branch configuration (upstream
configuration) when selecting a branch in the Repositories View.

A Properties View can be opened which will show the upstream
configuration for the selected branch. From there, it is possible to
open an editor for this configuration.

Alternatively, a right-click on the branch node will open the editor
directly.

Change-Id: I3976c90da855d7524cd9a64d765264e87622e36a
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>