Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

Works well, but only after the filter is once set to "Show all changes to selected resources and its children".

The reason is that the method createFollowFilterFor() of GitHistoryPage is only called from createFileWalker if all filter paths are file paths.

Patch-set: 4