Update patch set 9
Patch Set 9:

The main idea of this change is to dispose old synchronization data when new synchronization process starts.

It seams that none of eclipse views are disposed when they are cosed. Try changing something in "Display" view or in "History" view, then close it and open again, as you can see all modification are preserved.

Patch-set: 9