[DRAFT] Load history incrementally
Loading history incrementally can drastically reduce memory consumption
and the time needed to render the part of the history the user is
interested in. Most often only the most recent part of the history is
needed. Hence load only one batch of 256 commits initially and keep the
GenerateHistoryJob in order to incrementally load the next batch when
the user has scrolled down half of the loading batch size (128 commits).

Change-Id: I99c34657e6c5485a1303a13c4126b0049ea350d9
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>