Update patch set 2
Patch Set 2:

good question :) i have written a very basic IDE integration on top of EGit/JGit just for our company, which uses a CloneOperation to do the cloning. This stopped working, and it seemed like a bug to me. If it isn't at least the JavaDoc has to be adapted (it says pass null for no branch). After all this is "public" API.

Patch-set: 2