Remove braces from one-line statements
In I9e998aa40f50bfa13a1d9887d267e0ecc9833ec9, a Save Actions was added
to do this automatically on save. Because of existing code which
violated the style, this resulted in unrelated changes whenever one
edited a file.

So by cleaning this up once, there should be no further unrelated
changes in the future.

Change-Id: I1279030286ef177f3360ff70660a9365e4f3119a
Signed-off-by: Robin Stocker <robin@nibor.org>