Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(9 inline comments)

I found a number of places that should be blocks. I commented on the first few files and did not bother to look at the rest because the rule to not use blocks for single statements is for readability. The rule is not meant to be taken literally according to Java syntax rules. The rule is there to keep the code as easy to parse for humans while at the same time remove unnecessary clutter. Braces around single statements are *usually* clutter, just like extra parentheses *usuall* does not enhance readability.

We could have a lot of save rules, but we don't because the save rules area applied without thiking.


Some exceptions to this rule

Deeply nested "single-statements". Deeply for me is three levels or more where at least the outer level should have braces. 

Multiline statements can have braces without this being clutter.

Whatever may be ambigous to a human reader.

Please revert the save action and resubmit this patch with only the sane removals of braces.

Patch-set: 1