Update patch set 1
Patch Set 1:

I also didn't like the results in some places. That's the bad thing with the save actions, it's an all or nothing approach. Ok, I'll submit a new change to revert the save actions in this case.

You know, what bugs me about the whole situation is that the "with braces, except for some situations where there's only one short statement" is that it is kind of an unusual style and comes up in reviews a lot, making manual correction necessary. At work we use braces everywhere and make heavy use of save actions (formatting everything). And it's great because our codebase is as clean and unified as it gets with regard to coding style, and it's never an issue in reviews. But I realize that the recommended style won't be changed.

These exceptions you mentioned, would it be ok if I added a section about them here?: http://wiki.eclipse.org/EGit/Contributor_Guide#Coding_standards

Patch-set: 1