Don't cast node object to Ref
The reset command currently fails when a repository
is selected because the getLatestCommit method
casts the node's object to a Ref without checking the
type.

This change just passed the target branch and repository
to the getLatestCommit method for looking up the commit
that the target branch is pointing to.

Change-Id: Ic02ecb3e6cf7c75504a380d8088789029a9c0f4a