Update patch set 2
Patch Set 2:

But the job will still be scheduled the next time the index, refs, or resource event fires so presumably it will retry then and possibly succeed.

I'm not sure explicit retry logic is the right move unless there are explicit exception cases that can be recovered from.

Patch-set: 2