Use correct path when comparing with previous revision
Previously a RenameCallback was registered on the FollowFilter
to find the previous path of the file being compared.  This
led to incorrect behavior since the callback fires for all 
renames, not just for the commits processed through directly
calling RevWalk.next().

Now a RenameDetector is directly used after computing the
diffs between the head commit and the previous commit.
The path is only updated if the rename detector detects
that the current path was renamed between the two commits.

Change-Id: I562b7b3d0d96f6ce7294a9e2c84d9bb292ae6cf4