Fire config change event when edited via preference page
The previous behavior can be seen by opening the Configuration
preference page while the Properties view is being displayed
for a repository.  Selecting Apply or OK from the preference
page does not update the Properties view.  This is because
the config used by the preference page does not propagate the
change event to the repository's listeners list.

The preference page now registers a listener on the local
config being modified and passed the event to the repository
when fired.

Change-Id: Id0d2b13139eecb382c439c4b253131107ec75470