[repoView] Show commit each reference points to
Previously the full SHA-1 was displayed after the
name of each item under the 'References' node.

Now the abbreviated SHA-1 is shown followed by
the short message of the commit that SHA-1 points
to.

Also now show the symbolic name in braces using
the decorations styler instead of the qualifier
styler.

Change-Id: Iedf8a38d56d65d1eb33ea3fff0c8c9c615341839