[historyView] Do not store the RevWalk as a field in GitHistoryPage
When the history is redrawn a new RevWalk is created and configured
corresponding to the settings of the history view. This RevWalk is used
in the scheduled GenerateHistoryJob.

The RevWalk was stored in a field in the GitHistoryPage, although it is
not a property of such a Page. This can potentially lead to problems,
and indeed it caused bug 374675 which was fixed by
3604b5cd5fbf6ac354bc2a247115df7eade81959.

By passing it as a parameter the scope of the RevWalk instance is made
clearer.

Change-Id: I340f2a3cc7340dec781c09b8b8ced6a3fb03bdc5
Signed-off-by: Stefan Lay <stefan.lay@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>