Add support for Eclipse-SourceReferences header
Add extensions required to import a plug-in using SCM URL from its
MANIFEST.MF. To give it a try, open Plug-ins view, select a plug-in,
right click and Import As > Project from a Repository.... Then follow
instructions in the wizard. Importing with a tag is not supported yet,
see bug 367712.

This allows EGit compilation against all supported eclipse platform
versions but only builds the p2 repository when built against Juno.
This should be improved in a subsequent change.

Bug: 327381
Change-Id: Id9d721ae574a83ad9440a93ddb143b1271fc2b2b
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>