Define target platforms for the build
This simplifies setting up the Eclipse workspace and allows to use
Eclipse versions for development which are not matching the target
platform version. In order to set a target platform open it using the
target editor and click "Set as Target Platform". The EGit Maven build
now also is based on the same target platform definition. This way
dependencies used in the Maven build are identical to those used in
Eclipse.

This change is a prerequisite for change
Id9d721ae574a83ad9440a93ddb143b1271fc2b2b as this is using a feature of
the team framework only available in Juno. Without a target platform
definition this could not be built in older Eclipse versions.

JGit-Dependency: I36f2add2e10e5aac46bdf8e38aa799a690c63226
Change-Id: I106c6800ba849d6761daa6e453cf016c7ef8ecce
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>