Update patch set 16
Patch Set 16: Verified; Looks good to me, approved; IP review completed

The change looks good! I verified that the fix works against the latest 3.8 RC2 build (I20120521-2100) and the latest 4.2 RC2 build (I20120521-1900). I also verified that for targets < 3.8 the feature is disabled and EGit works as expected.

Given we are at RC2 already, and the test issues are not strictly related to this bug, I suggest we make an exception and merge this change, so that a wider audience can also test it. In the meantime we can continue to work on the test infrastructure.

Patch-set: 16