RFC Git API does not declare GitAPIException call()
Anyone that calls a Git API should catch GitAPIException and not just
the currently known exceptions.

During the process of cleaning this up a number of cases, but not all
cases where only checked exceptions were called into catching all
Exceptions except Error.

This changes adapt EGit to changes made in JGit for this reason.

Bug: 366914
JGit-Change-Id: I50380f13fc82c22d0036f47c7859cc3a77e767c5
Change-Id: I4de4902003a9f306912475e8b15b1e72d43e7475
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>