Update patch set 14
Patch Set 14:

After some more testing in Linux, I found out that the sequence of commits traversed by the RevWalk that determines the parent commits appears to be stable in Windows, but not in Linux. On Windows, this was hiding a bug in RepositoryActionHandler. After fixing this bug and forcing a sorting by date in the CommitSelectDialog, I got the two tests stable, at least in my Win/Linux environments.

Patch-set: 14