[historyView] Fix hover over Ref tags
After 5c5d942, the implementation of the hover must be adjusted,
otherwise the hover over the Ref tags (showing the full names of the
Ref) will wrongly activated when the mouse is width of first column left
of the actual tag.

This also sets the foreground color explicitly, as the current
implementation is rendered nearly unreadable on Ubuntu 12.04 (dark grey
on black).

Change-Id: I7feeed79fe1c996450ab8402615dcefddf85195b