[historyView] Fix label coordinates for hover detection
The coordinates which are used to detect whether a hover tooltip should
be shown didn't consider the width graph at the left side (cellX).

With only one lane, it's not a big deal, but the more lanes there were
on the left the more off the positions were.

Also widen the coordinates a bit to account for border.

Change-Id: I8fe269af7a73274ee6cbca96778ba7be0754d6bb
Signed-off-by: Robin Stocker <robin@nibor.org>