Update patch set 1
Patch Set 1:

Yes, that's it. The positions are off by the width of the graph. Try a ref at the far right when there are 5 lanes or so (jgit repo), and you will notice the tooltip doesn't appear. Now go a ref to the left and the tooltip text will be shown for the other ref because it's that much off.

With this change applied, it should work flawlessly.

Patch-set: 1