Update patch set 2
Patch Set 2: Verified; Looks good to me, but someone else must approve; IP review completed

This probably fixes the problem. I couldn't reprodue this eiher, even without this patch. Threading problems. 

A wonder if the clear call is needed. It was added, allegedly to fix a memory leak, but i'm not sure the
clear call belongs in the dispose method. For lack of a better solution, this one should go into 2.0

Patch-set: 2