Allow amending a published commit
Provide a message type from CommitMessageComponent so
calling classes know the severity of the message and
report amending a published commit as a warning.

A warning message is sufficient to notify the committer
that they may be rewriting published history but it
still should be allowed since there are valid cases
for amending published commits and then force pushing.

Change-Id: Ic2d4e279feb70c7f9ad45328cc369cdd981f3b96