[stagingView] Don't show other warnings when commit is not allowed
If a commit is not allowed (e.g. in conflicts state), the author and
committer text fields are set to empty. This lead to the status of the
commit message component to be not OK, and a warning was shown: "Invalid
committer specified".

Instead of the warning about the committer, we now show the message why
committing is not allowed at the moment. This message can be further
improved later to help the user, e.g. "Please resolve conflicts and then
commit".

Change-Id: I54ba78ee11d375cddf4d8475059153781b8380a3
Signed-off-by: Robin Stocker <robin@nibor.org>