Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The fix makes sure that no colors are disposed (which by itself is a good thing) but it does not fix the underlying problem that one job continues to draw even though it got cancelled. It will just continue to draw only to be replaced by the already scheduled new job.

Besides that, the changes are quite big for RC3/RC4. I would only submit them to 'master' together with improved job cancelling.

For 2.0 I would simply abort drawing when one of the colors is disposed. This also saves CPU cycles.

Possible fix/workaround) in CommitGraphTable.doPaint(Event):
if (c instanceof SWTCommit) {
  SWTLane lane = ((SWTCommit)c).getLane();
  if (lane !=null && lane.color != null
      && lane.color.isDisposed())
    return;
  }
}

I can provide the change, if others agree with this.

Patch-set: 1