Add clean action for repositories
This adds a "Clean..." menu item to the repository node in the
repository view. This has the advantage over the last try to add a clean
menu, that one can clean files and directories which are not part of the
workspace.

CQ: 6984
JGit-Dependency: Ibed0459005a5e306c010b9932f5b5fd107fb5448
Change-Id: I5136afd3333c6710b0c67c49618df6217e0d1fcf
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>