Move staging/unstaging into a background thread.
On large repositories using the staging view seems to freeze the UI
thread. This is "new", as previously one could start writing commit
messages while waiting for the staging to complete.

To restore the old behavior, the staging/unstaging is moved to a
background thread, disabling tables in the meantime to prevent the user
from triggering again before the first operation has finished.

Change-Id: I83dcf5be0dbee171fd177a6c25fd521c61ae016c