Add a "Push to Gerrit" action
This action allows to directly push to Gerrit by selecting a URI and the
name of the Gerrit branch.

Similar to the "Fetch from Gerrit" wizard, the user must select a URI
and the name of the Branch. The values are remembered for ease of use.

This is particularly helpful when working with several Gerrit branches
in parallel.

The action is added to the Repositories View menu as well as to the Team
-> Remote sub-menu.

Change-Id: Ib117aa0285f006a4c4bd7a54707b9ca6ae4cd475
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>