[repoView] Add decoration preferences for repositories
Allows the dirty flag text to be shown for repositories
displayed as well as decorating the repository icon based
upon the status of files in the repositories.

Allows quick scanning of the repository list to know which
ones have local changes and which ones do not.

Currently only the dirty flag is supported for text
decorations since putting the branch and branch status
as variables will force them to be displayed after the
repository path text which would regress the current
behavior.

Change-Id: Ie3d06c394f1ae1276eb358cd03290aa5516391d7