[historyView] Highlight filtered files in file diff viewer
The fix for bug 360073 changed the History view file diff viewer to show
all files of a commit, regardless of the resource filter. This made it
sometimes hard to find the files that the filter is focused on.

This change improves that by rendering the files that are not matched by
the filter in gray. The matching files are rendered in the same color as
before. With this, the files can be found at a glance instead of reading
through the paths.

Bug: 393610
JGit-Dependency: Icf911fe6fca131b2567514f54d66636a44561af1
Change-Id: I03da4b38d1591495cb290909f0e4c6e52270e97f
Signed-off-by: Robin Stocker <robin@nibor.org>