Only calculate containing resources for directories with .project
It should be save to do expensive checks on whether the given directory
belongs to a shared project only in the root of a project (determined by
looking for .project).

If it is, the iterator will be converted, if not, the expensive checks
will not be done for each subdirectory.

Change-Id: I8cd4dbb2240fa83d752e390604ffa235431c53e1