Rename "EGit Import Support" feature to avoid confusion
The old name is pretty generic and confuses people, resulting in "EGit
cannot be installed" questions on Bugzilla, Stackoverflow and to me
personally. They were checking all EGit features and were confused when
the installation did not complete, e.g. on Eclipse 3.7.

With the more specific name and an extended description, it should be
clear that this is an optional feature that is only needed by plugin
developers.

Change-Id: Ie7a137f202dba07e252fc4e587a0f78e66a1b5f7
Signed-off-by: Robin Stocker <robin@nibor.org>