[stagingView] Only offer actions applicable for all items
The action is run on all selected staging entries. So the available
actions for the whole selection should be the intersection of the
available actions of each entry.

This is a preconditions for the next change which will introduce a
Delete action, which should only be offered if all selected entries can
be deleted.

Bug: 384609
Change-Id: Ic0941b07f30af0c62ebc4860bf993ea51fa8c046
Signed-off-by: Robin Stocker <robin@nibor.org>