Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I would not use "PDE" in the name and also not use "Eclipse-SourceReferences" as this is an implementation detail which is not known by most users.

How about:

EGit Plug-in Import Support

Optional support for importing plug-ins and fragments from Git repositories.\n\
Requires Eclipse 3.8 or newer.\n


In addition, the name change should also be made in the plug-in 'org.eclipse.egit.import'.

Patch-set: 2