Don't interrupt user on first keystroke
The Push Commit... dialog always interrupted the user after the first
keystroke in the Target Ref Name field. To avoid this, we pre-fetch the
remote refs in the background.

The pushMode is now passed in the constructor of
RefContentAssistProvider. This fixes a latent bug in the caching story
of getRefsForContentAssist(boolean): Caches didn't distinguish
pushModes.

Bug: 381862
Change-Id: Ibbd3d03888312c5e268d00db8f654e742ab26787
Signed-off-by: Markus Keller <markus_keller@ch.ibm.com>