Update patch set 3
Patch Set 3:

There are two problems:

1) fetchJob.getResult() != Status.OK_STATUS
This condition is wrong: if the job ends with an error, then it will report an error on each key stroke (log and maybe also a dialog).

2) When the job got cancelled (for whatever reason) we should probably scheduled it again.


I know you hate it, but if you could add your credentials to each file (or add "and others."), it would save me some cycles.

Patch-set: 3