[repoView] Add Delete command for files and folders
Works on workspace and non-workspace resources.

For sharing the code with the Staging View, extracted the UI logic into
a new DeletePathsOperationUI class. This dispatches either to the
standard action from org.eclipse.ui or to DeletePathsOperation.

DeleteResourcesOperation was renamed to DeletePathsOperation to make it
clear that it operates on paths, not on potentially non-existing
resources.

Bug: 385001
Change-Id: I0d113cfa0fdc2c58ec33fa1c7ffd34b74e6c11a9
Signed-off-by: Robin Stocker <robin@nibor.org>