Update patch set 1
Patch Set 1: Do not submit

The code location for the change is correct, but the change restricts the fileset to projects which are imported into the workspace. If one changes a file which is not imported (e.g. .git/config), then that file would not be reported when it's dirty.

Patch-set: 1