Show problem decorations in staging view and commit dialog
This makes errors/warnings on the files visible while committing.
Before, these could only be noticed in other places before committing
(e.g. the package explorer).

In an earlier version of this change, IDecoratorManager was used. The
problem with that was that it could be turned off in the preferences and
that updating behavior was not correct.

So in this version, a plain ILabelDecorator is used and the decorations
are updated using a IResourceChangeListener.

Bug: 382755
Change-Id: I142a5f20b35f98675b90170dc1328ea637468c06
Signed-off-by: Robin Stocker <robin@nibor.org>