Show problem decorations in staging view and commit dialog
This makes errors/warnings on the files visible while committing.
Before, these could only be noticed in other places before committing
(e.g. the package explorer).

In an early draft of this change, the decoration was done directly in
StagingViewLabelProvider (without using IDecoratorManager). The problem
with that approach was that the decorations didn't get updated after
editing a file.

With the extension point and IDecoratorManager, updating the decoration
works. And, except for having to exclude GitLightweightDecorator via
IDecorationContext, it feels cleaner.

Bug: 382755
Change-Id: I142a5f20b35f98675b90170dc1328ea637468c06
Signed-off-by: Robin Stocker <robin@nibor.org>