Allow empty commit in the case of a merge commit
It sometimes happens that a resolved merge/cherry-pick is the same as
HEAD, in which case no files are staged for commit.

Before this change, the user was stuck in this state as the staging view
and commit dialog both refused to do a commit without staged files.

With this change, an empty commit is now allowed when the repository
state is MERGING_RESOLVED or CHERRY_PICKING_RESOLVED.

Bug: 385924
Change-Id: I6b9f5a91807726e24b5f1f9a5e8f841ed60e9f9c
Signed-off-by: Robin Stocker <robin@nibor.org>