Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

The patch works, but the part which tries to mimic the existing 'SaveFilter' is hard to understand.

I would do the following:

1. File a bug against Platform UI to make 'SaveFilter' API
2. For now, copy the whole class and mark it as such
3. Our code will then be much simpler: we create a subclass
   of 'SaveFilter', that tries file buffers, when
   super.select(...) returns 'false'

FYI: In case you are testing with multi-page (e.g. PDE) editors on 4.2, it won't work, but that's not our fault, see bug 386566 for details.

Patch-set: 2