Update patch set 1
Patch Set 1:

I don't know if it's a good idea to reuse the global Ctrl+# shortcut.

How would it work? When the Staging view is focused, it would cause a different kind of action to happen than when it is not focused. I think that may cause confusion (e.g. I wanted to have the Commit Dialog open but something else happened).

The reason for choosing Ctrl+Enter is that it's a common shortcut across multiple applications meaning "submit the current form". E.g. SourceTree also provides the same shortcut for committing, various mail applications for sending the message, etc. Ctrl+Enter was also the suggestion of the original bug reporter.

As a side note, the shortcut should be Command+Enter on OS X, I will correct that.

(By the way, with my keyboard layout, the necessary key sequence to press for the Commit dialog is this: Ctrl+Shift+3.)

Patch-set: 1