Update patch set 2
Patch Set 2:

> "change" and "7130" could use the folder icon

Good point. The only thing left is ordering, I find it cumbersome that "change" is now at the bottom of the branches list. Shouldn't it be sorted just like any other branch?

In other words, switching to hierarchical mode changes the ordering so leaves always precede non-leaf nodes. This is not right imho.

Patch-set: 2